# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlayMoney API",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "year": {
        "type": [
          "number",
          "null"
        ]
      },
      "month": {
        "type": [
          "number",
          "null"
        ]
      },
...
    },
    "parameters": {
...
      "year": {
        "schema": {
          "$ref": "#/components/schemas/year"
        },
        "required": false,
        "name": "year",
        "in": "path"
      },
      "month": {
        "schema": {
          "$ref": "#/components/schemas/month"
        },
        "required": false,
        "name": "month",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/leaderboard": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/year"
          },
          {
            "$ref": "#/components/parameters/month"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "topTraders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "total": {
                            "type": "number"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "userId",
                          "displayName",
                          "username",
                          "total",
                          "rank"
                        ]
                      }
                    },
                    "topCreators": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "total": {
                            "type": "number"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "userId",
                          "displayName",
                          "username",
                          "total",
                          "rank"
                        ]
                      }
                    },
                    "topPromoters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "total": {
                            "type": "number"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "userId",
                          "displayName",
                          "username",
                          "total",
                          "rank"
                        ]
                      }
                    },
                    "topQuesters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "total": {
                            "type": "number"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "userId",
                          "displayName",
                          "username",
                          "total",
                          "rank"
                        ]
                      }
                    },
                    "topReferrers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "total": {
                            "type": "number"
                          },
                          "rank": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "userId",
                          "displayName",
                          "username",
                          "total",
                          "rank"
                        ]
                      }
                    },
                    "userRankings": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "trader": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "displayName": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "string"
                                },
                                "avatarUrl": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "total": {
                                  "type": "number"
                                },
                                "rank": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "userId",
                                "displayName",
                                "username",
                                "total",
                                "rank"
                              ]
                            },
                            "creator": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "displayName": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "string"
                                },
                                "avatarUrl": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "total": {
                                  "type": "number"
                                },
                                "rank": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "userId",
                                "displayName",
                                "username",
                                "total",
                                "rank"
                              ]
                            },
                            "promoter": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "displayName": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "string"
                                },
                                "avatarUrl": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "total": {
                                  "type": "number"
                                },
                                "rank": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "userId",
                                "displayName",
                                "username",
                                "total",
                                "rank"
                              ]
                            },
                            "quester": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "displayName": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "string"
                                },
                                "avatarUrl": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "total": {
                                  "type": "number"
                                },
                                "rank": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "userId",
                                "displayName",
                                "username",
                                "total",
                                "rank"
                              ]
                            },
                            "referrer": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string"
                                },
                                "displayName": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "string"
                                },
                                "avatarUrl": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "total": {
                                  "type": "number"
                                },
                                "rank": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "userId",
                                "displayName",
                                "username",
                                "total",
                                "rank"
                              ]
                            }
                          }
                        },
                        {
                          "type": "null"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "topTraders",
                    "topCreators",
                    "topPromoters",
                    "topQuesters",
                    "topReferrers"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}